refactor(tools): move the check-pack family and shared workspace modules into tools/release
The pack-integrity family (check-pack.{ts,core.ts,test.ts}) and the
modules it shares (workspace.ts, types.ts) move verbatim into the
package — the boundary corollary: code the release pipeline owns lives
in one place. Consumer updates:

- workspaceRoot climbs two levels now that the module lives in
  <root>/tools/release instead of <root>/scripts
- vue-check's bin imports ../release/workspace.ts (same relative-import
  pattern it already used across scripts/)
- the root check:pack script points at the package entry, and the
  @pnpm/* devDependencies workspace.ts imports move to the package so
  resolution no longer leans on root hoisting
- //#check:pack turbo inputs over-approximate with tools/release/**
  since the entry left the root package's default-input scope

No script outside the moved set imports workspace.ts or types.ts
(verified: start-xvfb and upload-bundle-stats are self-contained), so
scripts/ needs no re-export shims. Zero behavior change.
