+++ tools/release/check-pack.core.ts
