+++ tools/release/package.json
@@ -12,6 +12,8 @@
     "typecheck": "tsc -p tsconfig.json --noEmit"
   },
   "devDependencies": {
+    "@pnpm/fs.find-packages": "catalog:",
+    "@pnpm/workspace.read-manifest": "catalog:",
     "@types/node": "catalog:",
     "typescript": "catalog:"
   }
