+++ turbo.json
@@ -82,7 +82,7 @@
       "passThroughEnv": ["CODECOV_TOKEN", "CI", "GITHUB_*"]
     },
     "//#check:pack": {
-      "inputs": ["$TURBO_DEFAULT$", "**/package.json"]
+      "inputs": ["$TURBO_DEFAULT$", "**/package.json", "tools/release/**"]
     },
     "//#test:root": {
       "inputs": ["scripts/**"]
